Mode the mode mapping into its own file
