+++ pylintrc
@@ -341,7 +341,7 @@ indent-string='  '
 max-line-length=100
 
 # Maximum number of lines in a module.
-max-module-lines=1000
+max-module-lines=1100
 
 # Allow the body of a class to be on the same line as the declaration if body
 # contains single statement.
